# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This repository is an example of an API made in Go with the To Do List theme.",
        "title": "To Do List API",
        "contact": {
            "name": "Lucas Santos",
            "url": "https://github.com/devlucassantos"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
        "/auth/signin": {
            "post": {
                "description": "Route that allows connecting the user to the system through their registration data. To connect a user it is necessary to inform the following data in the body of the request:\n|   Name   |  Type  |   Required  | Description\t\t|\n|----------|--------|-------------|-----------------|\n| email    | string |      x      | User email      |\n| password | string |      x      | User password   |",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "User Sign In",
                "operationId": "SignIn",
                "parameters": [
                    {
                        "description": "JSON responsible for sending all user sign in data to the server",
                        "name": "authJson",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SwaggerSignInRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User successfully signed in",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerAuthResponse"
                        }
                    },
                    "400": {
                        "description": "The user has made a bad request",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerValidationErrorResponse"
                        }
                    },
                    "401": {
                        "description": "The user is not authorized to access this account",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerUnauthorizedResponse"
                        }
                    },
                    "422": {
                        "description": "Some entered data could not be processed because it is not valid",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "An unexpected server error has occurred",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerGenericErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "description": "Route that allows you to register a user in the system. To register a user it is necessary to inform the following data in the body of the request:\n|   Name   |  Type  |   Required  | Description\t\t|\n|----------|--------|-------------|-----------------|\n| name     | string |      x      | Real user name  |\n| email    | string |      x      | User email      |\n| password | string |      x      | User password   |",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "User Sign Up",
                "operationId": "SignUp",
                "parameters": [
                    {
                        "description": "JSON responsible for sending all user registration data to the server",
                        "name": "authJson",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SwaggerSignUpRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User successfully registered",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerAuthResponse"
                        }
                    },
                    "400": {
                        "description": "The user has made a bad request",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerValidationErrorResponse"
                        }
                    },
                    "409": {
                        "description": "The user tried to register with the email of an existing user",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerConflictErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Some entered data could not be processed because it is not valid",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "An unexpected server error has occurred",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerGenericErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.SwaggerSignInRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "example@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "ex@mplePassw0rd"
                }
            }
        },
        "request.SwaggerSignUpRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "example@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "Example Name"
                },
                "password": {
                    "type": "string",
                    "example": "ex@mplePassw0rd"
                }
            }
        },
...
        "response.SwaggerAuthResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6Ikp..."
                },
                "email": {
                    "type": "string",
                    "example": "example@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Example Name"
                }
            }
        },
...
        "response.SwaggerConflictErrorResponse": {
            "type": "object",
            "properties": {
                "conflicts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Field example"
                    ]
                },
                "error_msg": {
                    "type": "string",
                    "example": "It is not possible to perform the operation because there are conflicting and/or duplicate data."
                }
            }
        },
...
        "response.SwaggerGenericErrorResponse": {
            "type": "object",
            "properties": {
                "error_msg": {
                    "type": "string",
                    "example": "Oops! An unexpected error has occurred."
                }
            }
        },
...
        "response.SwaggerInvalidField": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Description example"
                },
                "name": {
                    "type": "string",
                    "example": "Field example"
                }
            }
        },
...
        "response.SwaggerUnauthorizedResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Oops! You are not authorized."
                }
            }
        },
        "response.SwaggerValidationErrorResponse": {
            "type": "object",
            "properties": {
                "error_msg": {
                    "type": "string",
                    "example": "Some of the data entered is invalid."
                },
                "invalid_fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.SwaggerInvalidField"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}